Virtual service management via UI (WIP) + Multiple Access logs (#189)
Virtual service management via UI (WIP) + Multiple Access logs

---------

+++ ui/entrypoint.sh
@@ -7,6 +7,11 @@ if [ -z "$API_PROXY_PASS" ]; then
   exit 1
 fi
 
-envsubst '$API_PROXY_PASS' < /etc/nginx/conf.d/template.conf > /etc/nginx/conf.d/default.conf
+if [ -z "$GRPC_API_PROXY_PASS" ]; then
+  echo "ERROR: environment variable GRPC_API_PROXY_PASS not set"
+  exit 1
+fi
+
+envsubst '${API_PROXY_PASS},${GRPC_API_PROXY_PASS}' < /etc/nginx/conf.d/template.conf > /etc/nginx/conf.d/default.conf
 rm /etc/nginx/conf.d/template.conf
 nginx -g 'daemon off;'
